+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is modified from [Keep a Changelog](https://keepachangelog.com/en/1.0
 **Changed**
 - 時鐘遊戲更版，增加難易度順序，並修正時針位置
 - 連線遊戲支援平板遊玩
+- 更新網站標題
 
 ##### [2.3.5] - Educational Games
 **Changed**